Sequelize Database set up. Helperfunctions have not been made. Much of the psuedo code will have to be reomved. app.js runs database/index.js.
