+++ server/app.js
@@ -4,7 +4,7 @@ const musician = require('./musician');
 const band = require('./band');
 const listing = require('./listing');
 // Does not export anything yet. Just there to test the sequelize database.
-const sequelizeDatabase = require('../database/sequelize.config');
+const sequelizeDatabase = require('../database/index');
 require('dotenv').config();
 
 const app = express();
